Moving TransferManager test to integration tests
