+++ google-cloud-storage/src/main/java/com/google/cloud/storage/transfermanager/TransferManagerImpl.java
@@ -32,7 +32,7 @@ public final class TransferManagerImpl implements TransferManager {
   private final TransferManagerConfig transferManagerConfig;
   private final ExecutorService executor;
 
-  TransferManagerImpl(TransferManagerConfig transferManagerConfig) {
+  public TransferManagerImpl(TransferManagerConfig transferManagerConfig) {
     this.transferManagerConfig = transferManagerConfig;
     this.executor = Executors.newFixedThreadPool(transferManagerConfig.getMaxWorkers());
   }
